+++ packages/airnode-protocol/contracts/WIP/authorizers/admin/AirnodeFeeRegistryClient.sol
