+++ packages/airnode-protocol/contracts/WIP/authorizers/admin/AirnodeRequesterAuthorizerRegistryClient.sol
