+++ packages/airnode-protocol/contracts/WIP/authorizers/admin/AirnodeTokenPayment.sol
@@ -2,7 +2,7 @@
 pragma solidity 0.8.9;
 
 import "@openzeppelin/contracts/token/ERC20/extensions/IERC20Metadata.sol";
-import "../../authorizers/interfaces/IRequesterAuthorizerWithManager.sol";
+import "../../../authorizers/interfaces/IRequesterAuthorizerWithManager.sol";
 import "./AirnodeRequesterAuthorizerRegistryClient.sol";
 import "./AirnodeFeeRegistryClient.sol";
 import "./AirnodeTokenPaymentRolesWithManager.sol";
