+++ packages/airnode-protocol/contracts/WIP/authorizers/admin/AirnodeTokenPaymentRolesWithManager.sol
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.9;
 
-import "../../access-control-registry/RoleDeriver.sol";
-import "../../access-control-registry/AccessControlClient.sol";
-import "../../access-control-registry/interfaces/IAccessControlRegistry.sol";
+import "../../../access-control-registry/RoleDeriver.sol";
+import "../../../access-control-registry/AccessControlClient.sol";
+import "../../../access-control-registry/interfaces/IAccessControlRegistry.sol";
 import "./interfaces/IAirnodeTokenPaymentRolesWithManager.sol";
 
 /// @title Contract that implements generic AccessControlRegistry roles for
