+++ packages/airnode-protocol/contracts/WIP/authorizers/admin/interfaces/IAirnodeFeeRegistry.sol
