+++ packages/airnode-protocol/contracts/WIP/authorizers/admin/interfaces/IAirnodeFeeRegistryClient.sol
