+++ packages/airnode-protocol/contracts/WIP/authorizers/admin/interfaces/IAirnodeFeeRegistryRolesWithManager.sol
