+++ packages/airnode-protocol/contracts/WIP/authorizers/admin/interfaces/IAirnodeRequesterAuthorizerRegistry.sol
