+++ packages/airnode-protocol/contracts/WIP/authorizers/admin/interfaces/IAirnodeRequesterAuthorizerRegistryClient.sol
