+++ packages/airnode-protocol/contracts/WIP/authorizers/admin/interfaces/IAirnodeTokenLock.sol
