+++ packages/airnode-protocol/contracts/WIP/authorizers/admin/interfaces/IAirnodeTokenLockRolesWithManager.sol
