+++ packages/airnode-protocol/contracts/WIP/authorizers/admin/interfaces/IAirnodeTokenPayment.sol
