+++ packages/airnode-protocol/contracts/WIP/authorizers/admin/interfaces/IAirnodeTokenPaymentRolesWithManager.sol
