+++ packages/airnode-protocol/contracts/WIP/authorizers/admin/mock/MockApi3Token.sol
