+++ packages/airnode-protocol/contracts/WIP/authorizers/admin/mock/MockUsdcToken.sol
