+++ packages/airnode-protocol/test/WIP/authorizers/admin/AirnodeTokenPayment.sol.js
@@ -1,7 +1,7 @@
 /* globals context */
 const hre = require('hardhat');
 const { expect } = require('chai');
-const utils = require('../../utils');
+const utils = require('../../../utils');
 
 const chainId = 1;
 const oneDayInSeconds = 24 * 60 * 60;
